+++ pypeit/edgetrace.py
@@ -488,7 +488,14 @@ class EdgeTraceSet(DataContainer):
         self.log = None                 # Log of methods applied
         self.master_key = None          # Calibration key for master frame
         self.master_dir = None          # Directory for Master frames
-        self.maskdef_id = None          # Slit ID number from slit-mask design to record in SlitTraceSet
+        self.maskdef_id = None          # Slit ID number from slit-mask design matched to traced slits
+        self.align_slit = None          # Flag for alignment box slits from the slit-mask design
+        self.omodel_bspat = None        # Left edges predicted by the optical model (before x-correlation)
+        self.omodel_tspat = None        # Right edges predicted by the optical model (before x-correlation)
+        self.coeff_b = None             # Coefficients of the x-correlation between LEFT edges predicted
+                                        # by the optical model and traced on the image.
+        self.coeff_t = None             # Coefficients of the x-correlation between LEFT edges predicted
+                                        # by the optical model and traced on the image.
         self.maskdef_file = None        # File used to slurp in slit-mask design
 
     def _reinit_trace_data(self):
@@ -510,6 +517,11 @@ class EdgeTraceSet(DataContainer):
         self.design = None
         self.objects = None
         self.maskdef_id = None
+        self.align_slit = None
+        self.omodel_bspat = None
+        self.omodel_tspat = None
+        self.coeff_b = None
+        self.coeff_t = None
 
     @property
     def ntrace(self):
@@ -558,9 +570,9 @@ class EdgeTraceSet(DataContainer):
                     # table.Column(name='SLITRFOC', dtype=float, length=length,
                     #              description='Right edge of the slit in mm at the focal plane'),
                     table.Column(name='SLITLOPT', dtype=float, length=length,
-                                description='Left edge of the slit in pixel from optical model'),
+                                description='Left edge of the slit in pixel from optical model before x-correlation'),
                     table.Column(name='SLITROPT', dtype=float, length=length,
-                                description='Right edge of the slit in pixel from optical model'),
+                                description='Right edge of the slit in pixel from optical model before x-correlation'),
                     table.Column(name='SLITRA', dtype=float, length=length,
                                  description='Right ascension of the slit center (deg)'),
                     table.Column(name='SLITDEC', dtype=float, length=length,
@@ -795,19 +807,19 @@ class EdgeTraceSet(DataContainer):
             # synchronization to work correctly, we have to remove
             # fully masked traces. This is done inside sync().
 
-        # Left-right synchronize the traces
-        self.sync()
-        if show_stages:
-            self.show(title='After synchronizing left-right traces into slits')
-
-        # [DP] `maskdesign_matching` for now is only matching the traces found with the ones predicted by
-        # the slit-mask design. If we include the piece of code in which `maskdesign_matching` recovers
-        # missing traces, we should move the following 2 lines before `self.sync()`
+        # Match the traces found in the image with the ones predicted by
+        # the slit-mask design. If not expected traces are found in the image, they
+        # will be removed. If traces are missed, they will be added.
         if self.par['use_maskdesign']:
             msgs.info('-' * 50)
             msgs.info('{0:^50}'.format('Matching traces to the slit-mask design'))
             msgs.info('-' * 50)
             self.maskdesign_matching(debug=debug)
+
+        # Left-right synchronize the traces
+        self.sync()
+        if show_stages:
+            self.show(title='After synchronizing left-right traces into slits')
 
         # First manually remove some traces, just in case a user
         # wishes to manually place a trace nearby a trace that
@@ -839,8 +851,6 @@ class EdgeTraceSet(DataContainer):
 #        self.pca_refine(debug=debug)
 #        if show_stages:
 #            self.show()
-            
-        # TODO: Add mask_refine() when it's ready
 
         # Add this to the log
         self.log += [inspect.stack()[0][3]]
@@ -2533,6 +2543,10 @@ class EdgeTraceSet(DataContainer):
         if self.edge_fit is not None:
             self.edge_fit = self.edge_fit[:,keep]
         self.traceid = self.traceid[keep]
+        if self.maskdef_id is not None:
+            self.maskdef_id = self.maskdef_id[keep]
+        if self.align_slit is not None:
+            self.align_slit = self.align_slit[keep]
 
         if resort:
             # Resort by the spatial dimension
@@ -2724,6 +2738,10 @@ class EdgeTraceSet(DataContainer):
         self.edge_msk = self.edge_msk[:,srt]
         if self.edge_fit is not None:
             self.edge_fit = self.edge_fit[:,srt]
+        if self.maskdef_id is not None:
+            self.maskdef_id = self.maskdef_id[srt]
+        if self.align_slit is not None:
+            self.align_slit = self.align_slit[srt]
 
         # Reorder the trace numbers
         indx = self.traceid < 0
@@ -3940,6 +3958,10 @@ class EdgeTraceSet(DataContainer):
                                   np.zeros(_trace_cen.shape, dtype=float), axis=1)
         self.edge_msk = np.insert(self.edge_msk, loc, mask, axis=1)
         self.edge_fit = np.insert(self.edge_fit, loc, _trace_cen, axis=1)
+        if self.maskdef_id is not None:
+            self.maskdef_id = np.insert(self.maskdef_id, loc, -99)  # we don't know the slitmask id
+        if self.align_slit is not None:
+            self.align_slit = np.insert(self.align_slit, loc, 0)  # assuming it's not an alignment box
 
         if resort:
             self.spatial_sort()
@@ -4081,11 +4103,26 @@ class EdgeTraceSet(DataContainer):
                 omodel_bspat[i] = omodel_tspat[i]
                 omodel_tspat[i] = invert_order
 
+        # If there are overlapping slits, i.e., omodel_bspat[i+1]<omodel_tspat[i], switch those
+        for i in range(sortindx.size -1):
+            if omodel_tspat[sortindx][i] != -1 and omodel_bspat[sortindx][i+1] != -1 and \
+                    omodel_tspat[sortindx][i] > omodel_bspat[sortindx][i+1]:
+                tmt_bspat = omodel_tspat[sortindx][i]
+                omodel_tspat[sortindx[i]] = omodel_bspat[sortindx[i+1]]
+                omodel_bspat[sortindx[i+1]] = tmt_bspat
+                # Re-check If the `omodel_bspat` is greater than `omodel_tspat` and switch the order.
+                # It may happens if 3 slits are overlapping (true story!)
+                if omodel_bspat[sortindx[i]] > omodel_tspat[sortindx[i]]:
+                    invert_order = omodel_bspat[sortindx[i]]
+                    omodel_bspat[sortindx[i]] = omodel_tspat[sortindx[i]]
+                    omodel_tspat[sortindx[i]] = invert_order
+
         # This print a QA table with info on the slits (sorted from left to right) that fall in the current detector.
         # The only info provided here is `slitid`, which is called `dSlitId` in the DEIMOS design file. I had to remove
         # `slitindex` because not always matches `SlitName` from the DEIMOS design file.
         if not debug:
             num = 0
+            msgs.info('Expected slits on current detector')
             msgs.info('*' * 18)
             msgs.info('{0:^6s} {1:^12s}'.format('N.', 'dSlitId'))
             msgs.info('{0:^6s} {1:^12s}'.format('-' * 5, '-' * 9))
@@ -4099,6 +4136,7 @@ class EdgeTraceSet(DataContainer):
         # the IDL-based pipeline.
         if debug:
             num = 0
+            msgs.info('Expected slits on current detector')
             msgs.info('*' * 92)
             msgs.info('{0:^5s} {1:^10s} {2:^12s} {3:^12s} {4:^14s} {5:^16s} {6:^16s}'.format('N.',
                                                                                              'dSlitId', 'slitLen(mm)',
@@ -4109,7 +4147,7 @@ class EdgeTraceSet(DataContainer):
             msgs.info('{0:^5s} {1:^10s} {2:^12s} {3:^12s} {4:^14s} {5:^16s} {6:^14s}'.format('-' * 4, '-' * 9, '-' * 11,
                                                                                              '-' * 11, '-' * 13,
                                                                                              '-' * 18, '-' * 15))
-            for i in range(sortindx.shape[0]):
+            for i in range(sortindx.size):
                 if omodel_bspat[sortindx][i] != -1 or omodel_tspat[sortindx][i] != -1:
                     msgs.info('{0:^5d}{1:^14d} {2:^9.3f} {3:^12.3f} {4:^14.3f}    {5:^16.2f} {6:^14.2f}'
                               .format(num, self.spectrograph.slitmask.slitid[sortindx][i],
@@ -4181,8 +4219,12 @@ class EdgeTraceSet(DataContainer):
             plt.legend()
         msgs.info('SLIT_MATCH: RMS residuals for left and right edges: {}, {} pixels'.format(sigres_b, sigres_t))
 
-        bot_edge_pred = coeff_b[0] + coeff_b[1]*omodel_bspat if not switched else coeff_b[0] + coeff_b[1]*omodel_tspat
-        top_edge_pred = coeff_t[0] + coeff_t[1]*omodel_tspat if not switched else coeff_t[0] + coeff_t[1]*omodel_bspat
+        bot_edge_pred = omodel_bspat.copy()
+        bot_edge_pred[omodel_bspat!=-1] = coeff_b[0] + coeff_b[1] * omodel_bspat[omodel_bspat!=-1] if not switched else\
+                                          coeff_b[0] + coeff_b[1] * omodel_tspat[omodel_bspat!=-1]
+        top_edge_pred = omodel_tspat.copy()
+        top_edge_pred[omodel_tspat!=-1] = coeff_t[0] + coeff_t[1]*omodel_tspat[omodel_tspat!=-1] if not switched else\
+                                          coeff_t[0] + coeff_t[1]*omodel_bspat[omodel_tspat!=-1]
 
         # Find if there are missing traces.
         # Need exactly one occurrence of each index in "need"
@@ -4200,87 +4242,109 @@ class EdgeTraceSet(DataContainer):
         needadd_t[ind_t] = False
         needind_t = np.where(needadd_t)[0]  # edges we are missing
 
-        if (needind_b.shape[0] > 0) | (needind_t.shape[0] > 0):
+        if (needind_b.size > 0) | (needind_t.size > 0):
             msgs.warn('Missing edge traces: {} left and {} right'.format(needind_b.shape[0], needind_t.shape[0]))
-            msgs.warn('Some of them may have been removed by setting the parameter `minimum_slit_length`')
 
         if debug:
-            slitdesign_matching.plot_matches(self.edge_fit[:,self.is_left], ind_b, omodel_bspat, spat_bedge,
-                                             reference_row, self.spectrograph.slitmask.slitindx, nspat=self.nspat,
-                                             duplicates=dupl_b, missing=needind_b, edge='left')
-            slitdesign_matching.plot_matches(self.edge_fit[:,self.is_right], ind_t, omodel_tspat, spat_tedge,
-                                             reference_row, self.spectrograph.slitmask.slitindx, nspat=self.nspat,
-                                             duplicates=dupl_t, missing=needind_t, edge='right')
+            slitdesign_matching.plot_matches(self.edge_fit[:,self.is_left], ind_b, bot_edge_pred, reference_row,
+                                             self.spectrograph.slitmask.slitindx, nspat=self.nspat, duplicates=dupl_b,
+                                             missing=needind_b, edge='left')
+            slitdesign_matching.plot_matches(self.edge_fit[:,self.is_right], ind_t, top_edge_pred, reference_row,
+                                             self.spectrograph.slitmask.slitindx, nspat=self.nspat, duplicates=dupl_t,
+                                             missing=needind_t, edge='right')
 
-        # [DP] The code below is to add traces that are predicted but not found. For now we leave it commented, as we
-        # incorporate it in a second moment.
+        # Put duplicate flags for left and right traces together in one array
+        if np.any(dupl_b) or np.any(dupl_t):
+            edges_dupl = np.zeros(self.ntrace, dtype=bool)
+            if np.any(dupl_b):
+                edges_dupl[self.is_left] = dupl_b
+            if np.any(dupl_t):
+                edges_dupl[self.is_right] = dupl_t
 
-        # if needind_b.shape[0] > 0:
-        #     ind_b = np.append(ind_b, needind_b)
-        #     sortind_b = np.argsort(utils.index_of_x_eq_y(self.spectrograph.slitmask.slitindx[sortindx], ind_b, strict=True))
-        #     ind_b = ind_b[sortind_b]
-        #     lside = -np.ones(bot_edge_pred[needind_b].shape[0], dtype=int)
-        #     missing_left_traces = self.predict_traces(bot_edge_pred[needind_b], side=lside)
-        #     self.insert_traces(lside, missing_left_traces, mode='mask')
-        #
-        # if needind_t.shape[0] > 0:
-        #     ind_t = np.append(ind_t, needind_t)
-        #     sortind_t = np.argsort(utils.index_of_x_eq_y(self.spectrograph.slitmask.slitindx[sortindx], ind_t, strict=True))
-        #     ind_t = ind_t[sortind_t]
-        #     rside = np.ones(top_edge_pred[needind_t].shape[0], dtype=int)
-        #     missing_right_traces = self.predict_traces(top_edge_pred[needind_t], side=rside)
-        #     self.insert_traces(rside, missing_right_traces, mode='mask')
-        #
-        # if ((needind_b.shape[0] > 0)|(needind_t.shape[0] > 0))&(debug is True):
-        #     slitdesign_matching.plot_matches(self.edge_fit[:, self.is_left], ind_b, omodel_bspat,
-        #                  spat_bedge, reference_row, self.spectrograph.slitmask.slitindx, edge='left')
-        #     slitdesign_matching.plot_matches(self.edge_fit[:, self.is_right], ind_t, omodel_tspat,
-        #                  spat_tedge, reference_row, self.spectrograph.slitmask.slitindx, edge='right')
+            # Remove duplicate match
+            msgs.info('Removing duplicate matches: {} left and {} right'.format(ind_b[dupl_b].size, ind_t[dupl_t].size))
+            self.remove_traces(edges_dupl, rebuild_pca=True)
+            ind_b = ind_b[np.logical_not(dupl_b)]
+            ind_t = ind_t[np.logical_not(dupl_t)]
 
-        # [DP] The index resulting from the matching are provided separately for left and right edges (ind_b, ind_t)
-        # Therefore I can provide only one of the two here. If for a specific slit the right trace is found but not the
-        # left one, that slit will not have an associated id.
-        self.maskdef_id = self.spectrograph.slitmask.slitid[ind_b]
-        # In the eventuality (hopefully remote) that some matches are duplicates, some traces will not have
-        # a `maskdef_id` associated. We assign a value of `maskdef_id=-99` to those traces.
-        if np.any(dupl_b):
-            self.maskdef_id[dupl_b] = -99
+        #  The code below is to add traces that are predicted but not found.
+        if needind_b.size > 0:
+            msgs.info('Adding {} left missing edge(s)'.format(needind_b.size))
+            # Append the missing indices and re-sort all
+            ind_b = np.append(ind_b, needind_b)
+            sortind_b = np.argsort(utils.index_of_x_eq_y(self.spectrograph.slitmask.slitid[sortindx],
+                                                         self.spectrograph.slitmask.slitid[ind_b], strict=True))
+            ind_b = ind_b[sortind_b]
+            # define which side to add the trace and insert it
+            lside = -np.ones(bot_edge_pred[needind_b].shape[0], dtype=int)
+            missing_left_traces = self.predict_traces(bot_edge_pred[needind_b], side=lside)
+            self.insert_traces(lside, missing_left_traces, mode='mask')
 
-        # [DP] The following lines create two attributes, `design` and `object`, which store
-        # the matched slit-design and object information.
-        # Traced edges MUST be synchronized and NO duplicate matches should exist.
-        # TODO: pass the `design` and `object` to `EdgeTraceSet` and/or `SlitTraceSet` datamodel.
-        # TODO: [DP] There is some inconsistency for the two scripts below if there are duplicate matches.
-        #  For now I'll just put a condition, but it needs to be fixed.
-        if not np.any(dupl_b):
-            self._fill_design_table(ind_b, coeff_b, omodel_bspat, omodel_tspat)
-            self._fill_objects_table(ind_b)
+        if needind_t.size > 0:
+            msgs.info('Adding {} right missing edge(s)'.format(needind_t.size))
+            # Append the missing indices and re-sort all
+            ind_t = np.append(ind_t, needind_t)
+            sortind_t = np.argsort(utils.index_of_x_eq_y(self.spectrograph.slitmask.slitid[sortindx],
+                                                         self.spectrograph.slitmask.slitid[ind_t],strict=True))
+            ind_t = ind_t[sortind_t]
+            # define which side to add the trace and insert it
+            rside = np.ones(top_edge_pred[needind_t].shape[0], dtype=int)
+            missing_right_traces = self.predict_traces(top_edge_pred[needind_t], side=rside)
+            self.insert_traces(rside, missing_right_traces, mode='mask')
 
-    def _fill_design_table(self, ind, coeff, omodel_bspat, omodel_tspat):
+        if debug:
+            if (needind_b.size > 0) | np.any(dupl_b):
+                slitdesign_matching.plot_matches(self.edge_fit[:, self.is_left], ind_b, bot_edge_pred, reference_row,
+                                                 self.spectrograph.slitmask.slitindx, nspat=self.nspat, edge='left')
+            if (needind_t.size > 0) | np.any(dupl_t):
+                slitdesign_matching.plot_matches(self.edge_fit[:, self.is_right], ind_t, top_edge_pred, reference_row,
+                                                 self.spectrograph.slitmask.slitindx, nspat=self.nspat, edge='right')
+
+        self.maskdef_id = np.zeros(self.ntrace, dtype=int)
+        self.maskdef_id[self.is_left] = self.spectrograph.slitmask.slitid[ind_b]
+        self.maskdef_id[self.is_right] = self.spectrograph.slitmask.slitid[ind_t]
+
+        self.align_slit = np.zeros(self.ntrace, dtype=bool)
+        self.align_slit[self.is_left] = self.spectrograph.slitmask.alignment_slit[ind_b]
+        self.align_slit[self.is_right] = self.spectrograph.slitmask.alignment_slit[ind_t]
+
+        # Propagate the coefficients, `coeff_b` and `coeff_t`, of the x-correlation and the
+        # left and right spatial position of the slit edges from optical model (before x-correlation)
+        # with the purpose to fill a table with the information on slitmask design matching. The table will
+        # be filled out at the very end of the slit tracing process, in `get_slits()`.
+        self.coeff_b = coeff_b
+        self.coeff_t = coeff_t
+        self.omodel_bspat = omodel_bspat
+        self.omodel_tspat = omodel_tspat
+
+    def _fill_design_table(self, maskdef_id, coeff_b, coeff_t, omodel_bspat, omodel_tspat):
         """
         Fill :attr:`design` based on the results of the design
         registration.
 
         Args:
-            ind (:obj:`int`):
-                matched index for the slit-mask design data.
-            coeff (:obj:`numpy.array`):
+            maskdef_id (:obj:`numpy.array`):
+                Slit ID number from slit-mask design matched to traced slits.
+            coeff_b, coeff_t (:obj:`numpy.array`):
                 Fit parameters of the cross-correlation between slit-mask design
-                and traced edges.
-            omodel_bspat, omodel_tspat (:obj:`float`):
+                and traced edges for the left and right edges.
+            omodel_bspat, omodel_tspat (:obj:`numpy.array`):
                 Left and right spatial position of the slit edges from optical model
         """
         # Number of slits
-        nslits = len(ind)
+        nslits = self.nslits
         # Reference row
         reference_row = self.left_pca.reference_row if self.par['left_right_pca'] \
                             else self.pca.reference_row
+        # matched index for the slit-mask design data.
+        ind = utils.index_of_x_eq_y(self.spectrograph.slitmask.slitid, maskdef_id, strict=True)
+
         # Instantiate as an empty table
         self.design = EdgeTraceSet.empty_design_table(rows=nslits)
         # Save the fit parameters and the source file as table metadata
         self.design.meta['MASKFILE'] = None
-        self.design.meta['MASKOFF'] = coeff[0]
-        self.design.meta['MASKSCL'] = coeff[1]
+        self.design.meta['MASKOFF'] = coeff_b[0], coeff_t[0]
+        self.design.meta['MASKSCL'] = coeff_b[1], coeff_t[1]
         # Fill the columns
         self.design['TRACEID'] = np.arange(nslits, dtype=self.design['TRACEID'].dtype)
         self.design['TRACESROW'] = np.full(nslits, reference_row,
@@ -4300,14 +4364,14 @@ class EdgeTraceSet(DataContainer):
         self.design['ALIGN'] = self.spectrograph.slitmask.alignment_slit[ind].astype(
                                         dtype=self.design['ALIGN'].dtype)
 
-    def _fill_objects_table(self, ind):
+    def _fill_objects_table(self, maskdef_id):
         """
         Fill :attr:`objects` based on the result of the design
         registration.
 
         Args:
-            ind (:obj:`int`):
-                matched index for the slit-mask design data.
+            maskdef_id (:obj:`numpy.array`):
+                Slit ID number from slit-mask design matched to traced slits.
         """
         if self.spectrograph.slitmask.objects is None:
             # No object data available in slit mask design object
@@ -4317,8 +4381,7 @@ class EdgeTraceSet(DataContainer):
         # The index in the objects table are found by mapping the slit
         # index of each object in the design file to the slit index
         # included in the registration
-        obj_index = utils.index_of_x_eq_y(self.spectrograph.slitmask.slitindx, ind,
-                                          strict=True)
+        obj_index = utils.index_of_x_eq_y(self.spectrograph.slitmask.objects[:,0], maskdef_id, strict=True)
         # Number of objects
         nobj = len(obj_index)
         # Instantiate an empty table
@@ -4334,7 +4397,6 @@ class EdgeTraceSet(DataContainer):
         # in the original slit-mask design data
         self.objects['SLITINDX'] = utils.index_of_x_eq_y(self.objects['SLITID'],
                                                          self.design['SLITID'], strict=True)
-
 
 
 # NOTE: I'd like us to keep this commented mask_refine function around
@@ -4842,17 +4904,7 @@ class EdgeTraceSet(DataContainer):
         gpm = self.good_traces(include_box=True, good_orders=True)
         gpm = self.synced_selection(gpm, mode='neither')
 
-        # Initialize SlitTrace mask
-        slit_bitmask = slittrace.SlitTraceBitMask()
         nslit = int(np.sum(gpm)) // 2
-        slit_msk = np.zeros(nslit, dtype=slit_bitmask.minimum_dtype())
-        # Loop on left edges
-        for sidx, eidx in enumerate(np.where(gpm & self.is_left)[0]):
-            # Loop on SlitTrace mask keys
-            for key in slit_bitmask.keys():
-                if key in self.bitmask.keys() and np.all(self.bitmask.flagged(
-                        self.edge_msk[:, eidx], flag=key)):
-                    slit_msk[sidx] = slit_bitmask.turn_on(slit_msk[sidx], key)
 
         # Parse the data for the SlitTraceSet
         left = self.edge_fit[:,gpm & self.is_left]
@@ -4868,6 +4920,47 @@ class EdgeTraceSet(DataContainer):
             specmin = specmin[indx]/binspec
             specmax = specmax[indx]/binspec
 
+        if self.maskdef_id is not None:
+            # Check for mismatched `maskdef_id` in the left and right edges
+            mkd_id_mismatch = self.maskdef_id[self.is_left] != self.maskdef_id[self.is_right]
+            if np.any(mkd_id_mismatch):
+                msgs.warn("Mismatched `maskdefId` in left and right traces for {}/{} slits. ".format(
+                          self.maskdef_id[self.is_left][mkd_id_mismatch].size, self.nslits) +
+                          "Choosing the left edge `maskdefId` if it is not -99, otherwise choosing right one")
+            _maskdef_id = self.maskdef_id[gpm & self.is_left]
+            # choose the value of `maskdef_id` that is not -99.
+            mkd_id_bad = _maskdef_id == -99
+            # this may not work if the corresponding right edge is also -99. Assuming this is not the case
+            _maskdef_id[mkd_id_bad] = self.maskdef_id[gpm & self.is_right][mkd_id_bad]
+            if np.any(_maskdef_id == -99):
+                msgs.warn("Some slits don't have `maskdefId` assigned. Slit-mask design matching FAILED")
+                _maskdef_id = None
+                self.align_slit = None
+            else:
+                # Store the matched slit-design and object information in a table.
+                self._fill_design_table(_maskdef_id, self.coeff_b, self.coeff_t, self.omodel_bspat, self.omodel_tspat)
+                self._fill_objects_table(_maskdef_id)
+                _merged_designtab = table.join(self.design, self.objects, keys=['SLITID'])
+                _merged_designtab.sort('TRACEID')
+        else:
+            _maskdef_id = None
+            _merged_designtab = None
+
+        if self.align_slit is not None:
+            if np.any(self.align_slit):
+                self.edge_msk[:,self.align_slit] = self.bitmask.turn_on(self.edge_msk[:,self.align_slit], 'BOXSLIT')
+
+        # Initialize SlitTrace mask
+        slit_bitmask = slittrace.SlitTraceBitMask()
+        slit_msk = np.zeros(nslit, dtype=slit_bitmask.minimum_dtype())
+        # Loop on left edges
+        for sidx, eidx in enumerate(np.where(gpm & self.is_left)[0]):
+            # Loop on SlitTrace mask keys
+            for key in slit_bitmask.keys():
+                if key in self.bitmask.keys() and np.all(self.bitmask.flagged(
+                        self.edge_msk[:, eidx], flag=key)):
+                    slit_msk[sidx] = slit_bitmask.turn_on(slit_msk[sidx], key)
+
         # Instantiate and return
         return slittrace.SlitTraceSet(left, right, self.spectrograph.pypeline,
                                       det=self.traceimg.detector.det,
@@ -4875,7 +4968,7 @@ class EdgeTraceSet(DataContainer):
                                       PYP_SPEC=self.spectrograph.spectrograph, specmin=specmin,
                                       specmax=specmax, binspec=binspec, binspat=binspat,
                                       pad=self.par['pad'], mask_init=slit_msk,
-                                      maskdef_id=self.maskdef_id,
+                                      maskdef_id=_maskdef_id, maskdef_designtab=_merged_designtab,
                                       #maskdef_file=self.maskdef_file,  # ADD THIS BACK WHEN FITS ISSUE IS RESOLVED
                                       ech_order=ech_order)