+++ CHANGES.rst
@@ -37,6 +37,12 @@
 - The setup block is now a simple listing of the keywords and values
   used to identify the instrument configuration.
 - Refactor identify GUI and improve its docs
+- Pypeit now adds DEIMOS slits that are expected from the slitmask design
+  but not found in the tracing process.
+- PypeIt now flags as “BOXSLT” DEIMOS slits that are expected to be
+  alignment boxes from slitmask design.
+- Added a table with DEIMOS slitmask design and objects info to the
+  SlitTraceSet datamodel
 
 1.2.0 (15 Oct 2020)
 -------------------